+++ messaging/messaging_api.go
@@ -47,7 +47,7 @@ type Mux interface {
 
 	// Initializes new watcher which can start/stop watching on topic,
 	// eventually partition and offset.
-	NewPartitionWatcher(subscriberName string) ProtoManualWatcher
+	NewPartitionWatcher(subscriberName string) ProtoPartitionWatcher
 
 	// Disabled if the plugin config was not found.
 	Disabled() (disabled bool)
@@ -72,9 +72,9 @@ type ProtoWatcher interface {
 	StopWatch(topic string) error
 }
 
-// ProtoManualWatcher allows to subscribe for receiving of messages published
+// ProtoPartitionWatcher allows to subscribe for receiving of messages published
 // to selected topics, partitions and offsets
-type ProtoManualWatcher interface {
+type ProtoPartitionWatcher interface {
 	// WatchPartition starts consuming specific <partition> of a selected <topic>
 	// from a given <offset>. Offset is the oldest message index consumed,
 	// all previously published messages are ignored.