+++ CHANGELOG.md
@@ -32,8 +32,10 @@
 * kafka plugin initializes two multiplexers for dynamic mode (automatic partitions) and manual mode.
   Every multiplexer can create its own connection and provides access to different set of methods 
   (publishing to partition, watching on partition/offset)
-* ProtoWatcher from API was split to ProtoWatcher for dynamic mode and ProtoManualWatcher to listen on
-  specific partition/offset. Both of them allows only the respective watch methods   
+* ProtoWatcher from API was changed - methods WatchPartition and StopWatchPartition were removed 
+  from the ProtoWatcher interface and added to newly created ProtoPartitionWatcher. There is also a new 
+  method under Mux interface - NewPartitionWatcher(subscriber) which returns ProtoPartitionWatcher
+  instance that allows to call partition-related methods
 * fixes inside Mux.NewSyncPublisher() & Mux.NewAsyncPublisher() related to previous partition changes
 * TODO offset improvements
 * Known Issues: